requirements.txt
Flask
gunicorn
firebase-admin
python-dotenv
Pyrebase4
requests